ColumnarV1: widen ColDescriptor to 5×uint64_t, update all callers
All five ColDescriptor fields (type, null_offset, offsets_offset,
data_offset, data_size) promoted from uint32_t to uint64_t.
COLUMNAR_DESC_BYTES: 20 → 40.

Fixes ~3u alignment masks to use ~3ull to avoid upper-32-bit
truncation when cursor is uint64_t.

Updates all callers (buildColDescriptor return type, cursor variables
in ColumnBinaryOutputFormat, UserDefinedWebAssembly, and the wire
format gtest) to use uint64_t receivers.

Prepares the wire format for WASM memory64 where linear memory
can exceed 4 GiB.

+++ src/Functions/UserDefined/UserDefinedWebAssembly.cpp
@@ -462,7 +462,7 @@ public:
 
         // ── Build the columnar input buffer ──────────────────────────────────
         const uint32_t num_cols = static_cast<uint32_t>(cols.size());
-        uint32_t cursor = COLUMNAR_HEADER_BYTES + num_cols * COLUMNAR_DESC_BYTES;
+        uint64_t cursor = COLUMNAR_HEADER_BYTES + num_cols * COLUMNAR_DESC_BYTES;
 
         std::vector<ColDescriptor> descs(num_cols);
         std::vector<const IColumn *> inner_cols(num_cols);
@@ -492,7 +492,7 @@ public:
             cursor = buildColDescriptor(col, is_const, is_nullable, nrows, cursor, descs[ci]);
         }
 
-        uint32_t total_buf_size = cursor;
+        uint64_t total_buf_size = cursor;
 
         // ── Allocate buffer in WASM memory ───────────────────────────────────
         {
